+++ Bio/Phylo/TreeConstruction.py
@@ -577,7 +577,7 @@ class DistanceTreeConstructor(TreeConstructor):
 
     """
 
-    methods = ['nj', 'upgma']
+    methods = ['nj', 'upgma', 'wpgma']
 
     def __init__(self, distance_calculator=None, method="nj"):
         """Initialize the class."""
@@ -597,6 +597,8 @@ class DistanceTreeConstructor(TreeConstructor):
             tree = None
             if self.method == 'upgma':
                 tree = self.upgma(dm)
+            elif self.method == 'wpgma':
+                tree = self.wpgma(dm)
             else:
                 tree = self.nj(dm)
             return tree
@@ -604,10 +606,88 @@ class DistanceTreeConstructor(TreeConstructor):
             raise TypeError("Must provide a DistanceCalculator object.")
 
     def upgma(self, distance_matrix):
-        """Construct and return an UPGMA tree.
+        """Construct and return a UPGMA tree.
 
         Constructs and returns an Unweighted Pair Group Method
         with Arithmetic mean (UPGMA) tree.
+
+        :Parameters:
+            distance_matrix : DistanceMatrix
+                The distance matrix for tree construction.
+        """
+        if not isinstance(distance_matrix, DistanceMatrix):
+            raise TypeError("Must provide a DistanceMatrix object.")
+
+        # make a copy of the distance matrix to be used
+        dm = copy.deepcopy(distance_matrix)
+        # make a copy of the distance matrix to be used for counting
+        dm_count = copy.deepcopy(dm)
+        for i in range(1, len(dm_count)):
+            for j in range(0, i):
+                dm_count[i, j] = 1
+        # init terminal clades
+        clades = [BaseTree.Clade(None, name) for name in dm.names]
+        # init minimum index
+        min_i = 0
+        min_j = 0
+        inner_count = 0
+        while len(dm) > 1:
+            min_dist = dm[1, 0]
+            # find minimum index
+            for i in range(1, len(dm)):
+                for j in range(0, i):
+                    if min_dist >= dm[i, j]:
+                        min_dist = dm[i, j]
+                        min_i = i
+                        min_j = j
+
+            # create clade
+            clade1 = clades[min_i]
+            clade2 = clades[min_j]
+            inner_count += 1
+            inner_clade = BaseTree.Clade(None, "Inner" + str(inner_count))
+            inner_clade.clades.append(clade1)
+            inner_clade.clades.append(clade2)
+            # assign branch length
+            if clade1.is_terminal():
+                clade1.branch_length = min_dist * 1.0 / 2
+            else:
+                clade1.branch_length = min_dist * \
+                    1.0 / 2 - self._height_of(clade1)
+
+            if clade2.is_terminal():
+                clade2.branch_length = min_dist * 1.0 / 2
+            else:
+                clade2.branch_length = min_dist * \
+                    1.0 / 2 - self._height_of(clade2)
+
+            # update node list
+            clades[min_j] = inner_clade
+            del clades[min_i]
+
+            # rebuild distance matrix,
+            # set the distances of new node at the index of min_j
+            for k in range(0, len(dm)):
+                if k != min_i and k != min_j:
+                    dm[min_j, k] = ((dm[min_i, k] * dm_count[min_i, k]) + (dm[min_j, k] * dm_count[min_j, k])) * \
+                                   1.0 / (dm_count[min_i, k] + dm_count[min_j, k])
+
+                    dm_count[min_j, k] = dm_count[min_i, k] + dm_count[min_j, k]
+            dm_count.names[min_j] = "Inner" + str(inner_count)
+
+            del dm_count[min_i]
+
+            dm.names[min_j] = "Inner" + str(inner_count)
+
+            del dm[min_i]
+        inner_clade.branch_length = 0
+        return BaseTree.Tree(inner_clade)
+
+    def wpgma(self, distance_matrix):
+        """Construct and return a WPGMA tree.
+
+        Constructs and returns a Weighted Pair Group Method
+        with Arithmetic mean (WPGMA) tree.
 
         :Parameters:
             distance_matrix : DistanceMatrix
@@ -758,7 +838,11 @@ class DistanceTreeConstructor(TreeConstructor):
         if clade.is_terminal():
             height = clade.branch_length
         else:
-            height = height + max([self._height_of(c) for c in clade.clades])
+            for terminal in clade.get_terminals():
+                path = clade.get_path(target=terminal)
+                height = 0
+                for value in path:
+                    height += value.branch_length
         return height
 
 # #################### Tree Scoring and Searching Classes #####################